+++ NEWS.md
@@ -2,6 +2,11 @@
 
 * Bioconductor 3.16, October 2022
 
+## zellkonverter 1.7.5 (2022-09-13)
+
+* Minor changes for compatibility with **{cli}** v3.4.0
+  * Added tests for `verbose=TRUE` 
+
 ## zellkonverter 1.7.4 (2022-08-17)
 
 * Minor changes for compatibility with the upcoming **{Matrix}** 1.4-2 release
@@ -26,6 +31,11 @@
 # zellkonverter 1.6.0 (2022-04-27)
 
 * Bioconductor 3.15, April 2022
+
+## zellkonverter 1.6.5 (2022-09-13)
+
+* Minor changes for compatibility with **{cli}** v3.4.0
+  * Added tests for `verbose=TRUE` 
 
 ## zellkonverter 1.6.4 (2022-08-17)